+++ lib/core/config/router/routes/routes.dart
@@ -1,2 +1,3 @@
 export 'auth_routes.dart';
 export 'po_routes.dart';
+export 'admin_routes.dart';