+++ lib/core/core.dart
@@ -1,0 +1,3 @@
+export 'services/services.dart';
+export 'utils/utils.dart';
+export 'config/config.dart';