+++ lib/features/admin/domain/entities/admin_menu_option.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class MenuOptionEntity{
+  String titulo;
+  IconData icono;
+  List<String> subMenu;
+
+  MenuOptionEntity({
+    required this.titulo,
+    required this.icono,
+    required this.subMenu
+  });
+
+}