+++ lib/features/admin/presentation/widgets/admin_aside/sub_menu/expand_sub_menu.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class ExpandSubMenu extends StatelessWidget {
+
+  final bool isLast;
+
+  const ExpandSubMenu({
+    super.key,
+    required this.isLast
+  });
+
+  @override
+  Widget build(BuildContext context) {
+
+    return Expanded(
+      child: Container(
+        width: .5,
+        color: Colors.white.withValues(alpha: isLast ? 0 : 1),
+      ),
+    );
+
+  }
+
+}