+++ lib/features/features_screens.dart
@@ -3,3 +3,4 @@ export './shared/presentation/presentation.dart';
 export 'auth/presentation/presentation.dart';
 export 'point_of_sale/presentation/screens/screens.dart';
 export 'shared/presentation/screens/screens.dart';
+export 'admin/presentation/screens/screens.dart';