+++ lib/features/point_of_sale/presentation/screens/test_screen.dart
@@ -51,6 +51,7 @@ class TestScreen extends StatelessWidget {
       
     );
 
+    // ignore: dead_code
     return Container(
       color: Colors.white,
       child: SafeArea(